+++ lambda-fusion-authority/src/main/java/com/lambda/fusion/authority/client/controller/ClientController.java
@@ -1,11 +1,11 @@
 package com.lambda.fusion.authority.client.controller;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.lambda.cloud.core.utils.Assert;
 import com.lambda.fusion.authority.client.model.ClientEntity;
 import com.lambda.fusion.authority.client.model.ClientQuery;
 import com.lambda.fusion.authority.client.model.UpsertClient;
 import com.lambda.fusion.authority.client.service.ClientService;
+import com.lambda.fusion.authority.exception.AuthorityBusinessException;
 import com.lambda.fusion.core.utils.LoginUserUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -56,7 +56,9 @@ public class ClientController {
             @Parameter(description = "客户端编号", required = true) @PathVariable String id,
             @Parameter(description = "客户端信息", required = true) @RequestBody @Valid UpsertClient upsertClient) {
         ClientEntity original = clientService.getById(id);
-        Assert.notNull(original, "客户端不存在");
+        if (original == null) {
+            throw AuthorityBusinessException.clientNotFound(id);
+        }
         ClientEntity clientEntity = upsertClient.toEntity();
         clientEntity.setId(id);
         clientEntity.setSecret(original.getSecret());
@@ -68,7 +70,9 @@ public class ClientController {
     @Operation(summary = "删除客户端信息", description = "根据编号删除客户端信息")
     public void delete(@Parameter(description = "编号", required = true) @PathVariable String id) {
         ClientEntity client = clientService.getById(id);
-        Assert.notNull(client, "客户端不存在");
+        if (client == null) {
+            throw AuthorityBusinessException.clientNotFound(id);
+        }
         clientService.removeById(id);
     }
 }
